+++ src/deploy/missions/contracts/cross-chain/eth-portal.ts
@@ -1,9 +1,9 @@
 import { BaseDeployMission, TDeployArgs } from "@zero-tech/zdc";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { SignerWithAddress } from "@nomicfoundation/hardhat-ethers/signers";
-import { IZNSCampaignConfig, IZNSContracts } from "../../campaign/types";
-import { PORTAL_ROLE, ProxyKinds } from "../../constants";
-import { znsNames } from "./names";
+import { IZNSCampaignConfig, IZNSContracts } from "../../../campaign/types";
+import { PORTAL_ROLE, ProxyKinds } from "../../../constants";
+import { znsNames } from "../names";
 
 
 export class ZNSEthereumPortalDM extends BaseDeployMission<
