+++ src/deploy/missions/contracts/index.ts
@@ -1,11 +1,16 @@
-export * from "./address-resolver";
-export * from "./string-resolver";
-export * from "./registry";
-export * from "./root-registrar";
-export * from "./domain-token";
-export * from "./treasury";
-export * from "./curve-pricer";
-export * from "./access-controller";
-export * from "./meow-token/meow-token";
-export * from "./fixed-pricer";
-export * from "./sub-registrar";
+export * from "./zns-base/address-resolver";
+export * from "./zns-base/string-resolver";
+export * from "./zns-base/registry";
+export * from "./zns-base/root-registrar";
+export * from "./zns-base/domain-token";
+export * from "./zns-base/treasury";
+export * from "./zns-base/curve-pricer";
+export * from "./zns-base/access-controller";
+export * from "./zns-base/meow-token/meow-token";
+export * from "./zns-base/fixed-pricer";
+export * from "./zns-base/sub-registrar";
+
+export * from "./cross-chain/chain-resolver";
+export * from "./cross-chain/zchain-portal";
+export * from "./cross-chain/eth-portal";
+export * from "./cross-chain/zk-evm-bridge";
