+++ src/deploy/missions/contracts/zns-base/address-resolver.ts
@@ -1,12 +1,15 @@
-import { BaseDeployMission, TDeployArgs } from "@zero-tech/zdc";
+import {
+  BaseDeployMission,
+  TDeployArgs,
+} from "@zero-tech/zdc";
+import { ProxyKinds, ResolverTypes } from "../../../constants";
+import { znsNames } from "../names";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
+import { IZNSCampaignConfig, IZNSContracts } from "../../../campaign/types";
 import { SignerWithAddress } from "@nomicfoundation/hardhat-ethers/signers";
-import { IZNSCampaignConfig, IZNSContracts } from "../../campaign/types";
-import { ProxyKinds, ResolverTypes } from "../../constants";
-import { znsNames } from "./names";
 
 
-export class ZNSStringResolverDM extends BaseDeployMission<
+export class ZNSAddressResolverDM extends BaseDeployMission<
 HardhatRuntimeEnvironment,
 SignerWithAddress,
 IZNSCampaignConfig<SignerWithAddress>,
@@ -17,29 +20,26 @@ IZNSContracts
     kind: ProxyKinds.uups,
   };
 
-  contractName = znsNames.stringResolver.contract;
-  instanceName = znsNames.stringResolver.instance;
+  contractName = znsNames.addressResolver.contract;
+  instanceName = znsNames.addressResolver.instance;
 
   async deployArgs () : Promise<TDeployArgs> {
     const { accessController, registry } = this.campaign;
 
-    return [
-      await accessController.getAddress(),
-      await registry.getAddress(),
-    ];
+    return [await accessController.getAddress(), await registry.getAddress()];
   }
 
   async needsPostDeploy () {
     const {
       registry,
-      stringResolver,
+      addressResolver,
     } = this.campaign;
 
     const resolverInReg = await registry.getResolverType(
-      ResolverTypes.string,
+      ResolverTypes.address,
     );
 
-    const needs = resolverInReg !== await stringResolver.getAddress();
+    const needs = resolverInReg !== await addressResolver.getAddress();
     const msg = needs ? "needs" : "doesn't need";
 
     this.logger.debug(`${this.contractName} ${msg} post deploy sequence`);
@@ -50,15 +50,15 @@ IZNSContracts
   async postDeploy () {
     const {
       registry,
-      stringResolver,
+      addressResolver,
       config: {
         deployAdmin,
       },
     } = this.campaign;
 
     await registry.connect(deployAdmin).addResolverType(
-      ResolverTypes.string,
-      await stringResolver.getAddress(),
+      ResolverTypes.address,
+      await addressResolver.getAddress(),
     );
 
     this.logger.debug(`${this.contractName} post deploy sequence completed`);
