+++ src/deploy/missions/contracts/zns-base/fixed-pricer.ts
@@ -1,15 +1,14 @@
+import { ProxyKinds } from "../../../constants";
 import {
   BaseDeployMission,
   TDeployArgs,
 } from "@zero-tech/zdc";
-import { ProxyKinds } from "../../constants";
-import { znsNames } from "./names";
+import { znsNames } from "../names";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import { SignerWithAddress } from "@nomicfoundation/hardhat-ethers/signers";
-import { IZNSCampaignConfig, IZNSContracts } from "../../campaign/types";
+import { IZNSCampaignConfig, IZNSContracts } from "../../../campaign/types";
 
-
-export class ZNSCurvePricerDM extends BaseDeployMission<
+export class ZNSFixedPricerDM extends BaseDeployMission<
 HardhatRuntimeEnvironment,
 SignerWithAddress,
 IZNSCampaignConfig<SignerWithAddress>,
@@ -20,18 +19,15 @@ IZNSContracts
     kind: ProxyKinds.uups,
   };
 
-  contractName = znsNames.curvePricer.contract;
-  instanceName = znsNames.curvePricer.instance;
+  contractName = znsNames.fixedPricer.contract;
+  instanceName = znsNames.fixedPricer.instance;
 
   async deployArgs () : Promise<TDeployArgs> {
     const {
       accessController,
       registry,
-      config: {
-        rootPriceConfig,
-      },
     } = this.campaign;
 
-    return [await accessController.getAddress(), await registry.getAddress(), rootPriceConfig];
+    return [await accessController.getAddress(), await registry.getAddress()];
   }
 }
