+++ src/deploy/missions/contracts/zns-base/meow-token/mainnet-data.ts
